+++ gomoku_rust_lib/src/negamax.rs
@@ -1,16 +1,22 @@
 use crate::state::create_child;
 use crate::state::State;
 use std::cmp::Reverse;
-pub fn negamax(mut state: &mut State, depth: i32, mut alpha: i32, mut beta: i32, color: i8) -> i32 {
+pub fn negamax(
+    mut state: &mut State,
+    depth: isize,
+    mut alpha: isize,
+    mut beta: isize,
+    color: i8,
+) -> isize {
     if depth != 0 {
         state.available_move = create_child(&mut state);
         state.available_move.sort_by_key(|d| Reverse(d.heuristic));
     }
     // println!("current state: {:?} current heuristic {} depth {}", state.current_move, state.heuristic, depth);
     if depth == 0 || state.available_move.len() == 0 {
-        return state.heuristic * color as i32;
+        return state.heuristic * color as isize;
     }
-    let mut value: i32 = -1000;
+    let mut value: isize = -1000;
     let len = state.available_move.len();
     for child in 0..len {
         value = std::cmp::max(
@@ -33,7 +39,7 @@ pub fn negamax(mut state: &mut State, depth: i32, mut alpha: i32, mut beta: i32,
     return value;
 }
 
-pub fn return_move(state: &State, heuristic: i32) -> ((isize, isize), i32) {
+pub fn return_move(state: &State, heuristic: isize) -> ((isize, isize), isize) {
     // println!("heuristic ret = {:?}", heuristic);
     let len = state.available_move.len();
     for child in 0..len {
